feat: Creando nuevos archivos y configuraciones para poder hacer el crud de las card de cita

+++ app/src/main/java/com/example/vivasalud/data/local/dao/UserDao.kt
@@ -16,5 +16,5 @@ interface UserDao {
     @Delete
     fun delete(user: User)
     @Query("SELECT * FROM tb_user WHERE id = :id")
-    fun getUserById(id: Int): User?
+    fun getUserById(id: Long): User?
 }
